+++ lib/string.h
@@ -1,9 +1,10 @@
 #ifndef __LIB_STRING_H
 #define __LIB_STRING_H
 #include "stdint.h"
-void memset(void* dst_, uint8_t value, uint32_t size);
+#include "global.h"
+void memset(void *dst_, uint8_t value, uint32_t size);
 void memcpy(void* dst_, const void* src_, uint32_t size);
-int8_t memcmp(const void* a_, const void* b_, uint32_t size);
+int memcmp(const void* a_, const void* b_, uint32_t size);
 char* strcpy(char* dst_, const char* src_);
 uint32_t strlen(const char* str);
 int8_t strcmp (const char *a, const char *b);
